+++ sketch-view-core/src/androidTest/kotlin/com/github/panpf/sketch/view/core/test/request/ImageRequestViewExtensionsTest.kt
@@ -30,7 +30,7 @@ class ImageRequestViewExtensionsTest {
         val imageView1 = ImageView(context1)
         ImageRequest(imageView1, uri).apply {
             Assert.assertSame(context1, this.context)
-            Assert.assertEquals("asset://sample.jpeg", uri)
+            Assert.assertEquals("file:///android_asset/sample.jpeg", uri)
             Assert.assertNull(this.listener)
             Assert.assertNull(this.progressListener)
             Assert.assertEquals(ImageViewTarget(imageView1), this.target)